requirements.txt
Pillow
PyTurboJPEG
numpy
pyside6
pillow
